Check if client-cli is there after compile, and if not, spit out short lived token with warning

+++ scripts/docker/register.sh
@@ -44,7 +44,12 @@ then
     meson compile -C build
 fi
 
-HA_WS_URL=ws://localhost:8123/api/websocket HA_API_TOKEN="${TOKEN}" LD_LIBRARY_PATH=build/subprojects/curl-8.5.0/build/lib/.libs/ build/client-cli ha-get-token 1>longtoken.txt
-
-echo HA_WS_URL=ws://localhost:8123/api/websocket HA_API_TOKEN="`cat longtoken.txt`"
-rm longtoken.txt
+if [ -x "build/client-cli" ]
+then
+    HA_WS_URL=ws://localhost:8123/api/websocket HA_API_TOKEN="${TOKEN}" LD_LIBRARY_PATH=build/subprojects/curl-8.5.0/build/lib/.libs/ build/client-cli ha-get-token 1>longtoken.txt
+    echo HA_WS_URL=ws://localhost:8123/api/websocket HA_API_TOKEN="`cat longtoken.txt`"
+    rm longtoken.txt
+else
+    echo "WARNING - compile of client-cli failed, providing SHORT LIVED TOKEN"
+    echo HA_WS_URL=ws://localhost:8123/api/websocket HA_API_TOKEN="${TOKEN}"
+fi
